feat(pos): weight sale, quantity rules, scales auto-detect (v1.0.58)
Add PIECE/WEIGHT/SERVICE with DECIMAL quantities, POS weight modal,
X/Z shift returns, Electron scales module with COM auto-detect and picker.

+++ web-backend/src/main/java/com/pos/controller/ProductController.java
@@ -131,7 +131,7 @@ public class ProductController {
     @PreAuthorize("hasAnyRole('ADMIN', 'MANAGER')")
     public ResponseEntity<ProductResponse> adjustStock(
         @PathVariable UUID id,
-        @RequestParam int quantity,
+        @RequestParam BigDecimal quantity,
         @RequestParam(defaultValue = "ADJUSTMENT") String movementType,
         @RequestParam(required = false) String notes,
         @RequestParam(required = false) Integer storeId
