+++ web-backend/src/main/java/com/pos/dto/product/ProductLifecycleSummaryResponse.java
@@ -1,5 +1,6 @@
 package com.pos.dto.product;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -7,7 +8,7 @@ public record ProductLifecycleSummaryResponse(
     UUID productId,
     String sku,
     String name,
-    int currentStock,
+    BigDecimal currentStock,
     long stockDispatched,
     Instant productCreatedAt,
     long restockUnits,
