+++ web-backend/src/main/java/com/pos/dto/warehouse/StockInventoryResponse.java
@@ -1,5 +1,6 @@
 package com.pos.dto.warehouse;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
@@ -12,7 +13,7 @@ public record StockInventoryResponse(
     String status,
     String notes,
     int totalLines,
-    int totalDifference,
+    BigDecimal totalDifference,
     String createdByName,
     Instant createdAt,
     List<StockInventoryLineResponse> lines
